refactor: クリーンアーキテクチャに従い型定義をドメイン層に集約
- src/types/shopping.ts: ShoppingStatus, ShoppingItem, UpsertShoppingItemInput, PurchaseInput を追加
- src/types/barcode.ts: ProductInfo, BarcodeLookupSource, BarcodeLookupResult を追加
- src/types/user.ts: NotificationPreferences, UpdatePrefs を追加
- src/types/calendar.ts: PendingLotRemoval, computeCalendarDelta を追加
- src/types/item.ts: ItemFilters, ItemSortKey, ConsumeParams, ConsumeLotParams を追加
- src/lib/itemCache.ts: ItemSortKey のインポートをフック層からドメイン層（types/item.ts）に修正
- src/hooks/useCalendarConsume.ts: カレンダー操作のDB呼び出しをルートからフックへ抽出
- src/routes/_auth.calendar.tsx: プレゼンテーション層のSupabase直呼び出しを排除
